Move fold code to its own component
