+++ src/touchbar/components/TextEditor.js
@@ -1,19 +1,15 @@
 'use babel';
 
-/* eslint import/no-extraneous-dependencies: 0 */
-
 import React, { Component, Fragment } from 'react';
 import { nativeImage } from 'remote';
 import {
   hexToHsl,
-  getActiveElement,
   createOcticonImage,
   executeAtomCommand,
 } from '../../utils';
 
 import ToggleSidebarButton from './ToggleSidebarButton';
-
-const buildArray = (size) => Array.from(Array(size).keys());
+import FoldCodePopover from './FoldCodePopover';
 
 export default class TextEditor extends Component {
   constructor(props) {
@@ -37,28 +33,10 @@ export default class TextEditor extends Component {
     );
 
     this.buildOcticonIcons();
-
-    this.foldCode = this.foldCode.bind(this);
   }
 
   async buildOcticonIcons() {
     const whiteColor = '#ffffff';
-
-    const foldPromise = createOcticonImage({
-      icon: 'fold',
-      color: whiteColor,
-      height: 204,
-      width: 204,
-      scaleFactor: 10.0,
-    });
-
-    const unfoldPromise = createOcticonImage({
-      icon: 'unfold',
-      color: whiteColor,
-      height: 204,
-      width: 204,
-      scaleFactor: 10.0,
-    });
 
     const folderPromise = createOcticonImage({
       icon: 'file-directory',
@@ -69,52 +47,19 @@ export default class TextEditor extends Component {
     });
 
     const [
-      foldIcon,
-      unfoldIcon,
       folderIcon,
-    ] = await Promise.all([foldPromise, unfoldPromise, folderPromise]);
+    ] = await Promise.all([folderPromise]);
 
-    this.setState({ foldIcon, unfoldIcon, folderIcon });
-  }
-
-  foldCode(index) {
-    return () => {
-      const activeElement = getActiveElement();
-      atom.commands.dispatch(activeElement, `editor:fold-at-indent-level-${index}`);
-    };
-  }
-
-  renderFoldButtons() {
-    const foldingLevels = 5;
-    return buildArray(foldingLevels).map(index => (
-      <button
-        key={`fold-${index}`}
-        onClick={this.foldCode(index + 1)}
-        iconPosition="left"
-      >
-        {`At level ${index + 1}`}
-      </button>
-    ));
+    this.setState({ folderIcon });
   }
 
   render() {
-    const { foldIcon, unfoldIcon, folderIcon } = this.state;
+    const { folderIcon } = this.state;
 
     return (
       <Fragment>
         <ToggleSidebarButton />
-        <popover
-          icon={foldIcon}
-        >
-          <button
-            onClick={() => executeAtomCommand('editor:unfold-all')}
-            icon={unfoldIcon}
-            iconPosition="left"
-          >
-            Unfold code
-          </button>
-          {this.renderFoldButtons()}
-        </popover>
+        <FoldCodePopover />
         <button
           onClick={() => executeAtomCommand('command-palette:toggle')}
           icon={this.searchIcon}
